+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @ai-sdk/langchain
 
+## 3.0.0-canary.152
+
+### Patch Changes
+
+- ai@7.0.0-canary.152
+
+## 3.0.0-canary.151
+
+### Patch Changes
+
+- ai@7.0.0-canary.151
+
+## 3.0.0-canary.150
+
+### Patch Changes
+
+- ai@7.0.0-canary.150
+
+## 3.0.0-canary.149
+
+### Patch Changes
+
+- Updated dependencies [e3d9c0e]
+  - ai@7.0.0-canary.149
+
+## 3.0.0-canary.148
+
+### Patch Changes
+
+- Updated dependencies [2852a84]
+  - ai@7.0.0-canary.148
+
+## 3.0.0-canary.147
+
+### Patch Changes
+
+- ai@7.0.0-canary.147
+
 ## 3.0.0-canary.146
 
 ### Patch Changes